requirements.txt
# Root requirements (development convenience)
#
# Notes (Windows):
# - RDKit: recommended via conda, not pip:
#     conda install -c conda-forge rdkit
# - torch-geometric ecosystem (torch-scatter/torch-sparse/pyg):
#   install matching your PyTorch/CUDA version per official instructions:
#     https://pytorch-geometric.readthedocs.io/en/latest/install/installation.html

numpy>=1.21
pandas>=1.3
pyyaml>=6.0
easydict>=1.10
tqdm>=4.64

# ML / DL
torch>=2.0

# Graph learning (install method varies by platform)
torch-geometric>=2.4
torch-scatter>=2.1

# Data
lmdb>=1.4

# Bio / geometry
biopython>=1.81
scipy>=1.10

# Chemistry
joblib>=1.2

# LLM agents (optional)
aiohttp>=3.9

# Optional (only needed by some utilities)
# openbabel-wheel>=3.1.1.22
# autodocktools  # vina helper uses AutoDockTools path; usually installed separately

